requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
defusedxml==0.7.1
idna==3.10
jira==3.8.0
lxml==6.0.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
python-docx==1.2.0
python-jira==3.4.1
openai>=1.93.0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
typing_extensions==4.14.1
urllib3==2.5.0
